Change project files to SDK project files
Make necessary changes like
- only netstandard2.0 or net472 as target frameworks
- remove obsolete files
- update packages to recent versions
- adjust unit tests

+++ SharpMap.Data.Providers.GeoPackage/Tiles/GpkgTileSchema.cs
@@ -52,12 +52,12 @@ namespace SharpMap.Data.Providers.Tiles
             return _resolutions[levelId].Top; 
         }
 
-        public int GetMatrixWidth(string levelId)
+        public long GetMatrixWidth(string levelId)
         {
             return _resolutions[levelId].MatrixWidth;
         }
 
-        public int GetMatrixHeight(string levelId)
+        public long GetMatrixHeight(string levelId)
         {
             return _resolutions[levelId].MatrixHeight;
         }
